requirements.txt
faker
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
django-mssql-backend==2.4.2
pytz==2019.2
sqlparse==0.3.0
